requirements.txt
bandit>=1.0,<2.0
dj-database-url==0.4.1
dj-email-url==0.0.8
django-click==1.2.0
Django==1.8.14
django-cors-headers==1.1.0
djangorestframework==3.4.3
djorm-ext-pgfulltext==0.9.3
model-mommy==1.2.6
flake8==3.0.4
newrelic==2.68.0.50
psycopg2==2.6.2
pytest==2.9.2
pytest-cov==2.3.1
pytest-django==2.9.1
selenium==2.53.6
six==1.10.0
waitress==0.9.0
whitenoise==3.2.1
PyJWT==1.4.2
requests==2.11.0
python-dotenv==0.5.1
httmock==1.2.5
xlrd==1.0.0
django-debug-toolbar==1.5
